Better type for pointer button mask
This is a very limited bit field, so use an 8 bit type to clearly show
how many bits are available.

+++ common/rfb/CMsgWriter.h
@@ -59,7 +59,7 @@ namespace rfb {
     void writeFence(uint32_t flags, unsigned len, const uint8_t data[]);
 
     void writeKeyEvent(uint32_t keysym, uint32_t keycode, bool down);
-    void writePointerEvent(const core::Point& pos, int buttonMask);
+    void writePointerEvent(const core::Point& pos, uint8_t buttonMask);
 
     void writeClientCutText(const char* str);
